move IndexKernelUtils.h to folder comm: prepare upstream the cmake for sycl free func code
