+++ src/ATen/CMakeLists.txt
@@ -12,6 +12,7 @@ file(GLOB xpu_native_cpp "native/xpu/*.cpp" "native/sparse/*.cpp" "native/sparse
 file(GLOB xpu_sycl "native/xpu/sycl/*.cpp" "native/sparse/xpu/sycl/*.cpp" "native/nested/xpu/sycl/*.cpp" "native/transformers/sycl/*.cpp" "native/quantized/sycl/*.cpp")
 file(GLOB xpu_sycltla "native/transformers/xpu/flash_attn/sycltla/*.cpp")
 file(GLOB xpu_sleep "xpu/Sleep.cpp")
+file(GLOB xpu_comm "native/xpu/sycl/comm/*.cpp" )
 
 if(USE_ONEMKL_XPU)
   file(GLOB xpu_mkl "native/xpu/mkl/*.cpp")
@@ -20,11 +21,13 @@ endif()
 list(APPEND ATen_XPU_NATIVE_CPP_SRCS ${xpu_native_cpp})
 list(APPEND ATen_XPU_SYCL_SRCS ${xpu_sycl} ${xpu_sleep})
 list(APPEND ATen_XPU_SYCLTLA_SRCS ${xpu_sycltla})
+list(APPEND ATen_XPU_SYCL_COMM_SRCS ${xpu_comm})
 
 set(ATen_XPU_MKL_SRCS ${ATen_XPU_MKL_SRCS} PARENT_SCOPE)
 set(ATen_XPU_NATIVE_CPP_SRCS ${ATen_XPU_NATIVE_CPP_SRCS} PARENT_SCOPE)
 set(ATen_XPU_SYCL_SRCS ${ATen_XPU_SYCL_SRCS} PARENT_SCOPE)
 set(ATen_XPU_SYCLTLA_SRCS ${ATen_XPU_SYCLTLA_SRCS} PARENT_SCOPE)
+set(ATen_XPU_SYCL_COMM_SRCS ${ATen_XPU_SYCL_COMM_SRCS} PARENT_SCOPE)
 
 # ATen XPU headers
