+++ src/ATen/native/xpu/sycl/Indexing.h
@@ -11,7 +11,7 @@
 #pragma once
 
 #include <ATen/native/xpu/sycl/BatchKernel.h>
-#include <ATen/native/xpu/sycl/IndexKernelUtils.h>
+#include <ATen/native/xpu/sycl/comm/IndexKernelUtils.h>
 #include <ATen/native/xpu/sycl/Loops.h>
 #include <ATen/native/xpu/sycl/MemoryAccess.h>
 #include <comm/TensorInfo.h>
