+++ src/ATen/native/xpu/sycl/ScatterGatherKernels.cpp
@@ -21,7 +21,7 @@ DISABLE_RETURN_TYPE_WARNING_BEGIN
 #include <ATen/native/ScatterGatherChecks.h>
 #include <ATen/native/TensorIterator.h>
 #include <ATen/native/xpu/sycl/Atomics.h>
-#include <ATen/native/xpu/sycl/IndexKernelUtils.h>
+#include <ATen/native/xpu/sycl/comm/IndexKernelUtils.h>
 #include <ATen/native/xpu/sycl/MemoryAccess.h>
 #include <ATen/native/xpu/sycl/OffsetCalculator.h>
 #include <comm/SYCLContext.h>
