+++ src/CMakeLists.txt
@@ -12,6 +12,7 @@ set(ATen_XPU_NATIVE_CPP_SRCS)
 set(ATen_XPU_SYCL_SRCS)
 set(ATen_XPU_SYCLTLA_SRCS)
 set(ATen_XPU_XCCL_SRCS)
+set(ATen_XPU_SYCL_COMM_SRCS)
 
 set(ATen_XPU_INCLUDE_DIRS ${TORCH_XPU_OPS_ROOT}/src CACHE STRING "ATen XPU Include directory")
 set(ATen_XPU_FLASH_ATTN_DIRS
